prueba/requirements.txt
asgiref
certifi
charset-normalizer
Django
django-cors-headers
django-countries
djangorestframework
djangorestframework-jwt
djangorestframework-simplejwt
dropbox
idna
pillow
ply
PyJWT
python-dotenv
requests
six
sqlparse
stone
typing_extensions
tzdata
urllib3
